seperation of concerns
