+++ index.js
@@ -1,2 +1,8 @@
 const StacheEntryPlugin = require('./src/entry');
-module.exports = new StacheEntryPlugin();
+const cliCommands = require('./src/cli-commands/cli-commands');
+const entryPlugin = new StacheEntryPlugin();
+
+const publicApi = {};
+publicApi.preload = entryPlugin.preload;
+publicApi.runCommand = cliCommands.runCommand;
+module.exports = publicApi;
