+++ package.json
@@ -20,12 +20,12 @@
   },
   "homepage": "https://github.com/blackbaud/skyux-builder-plugin-stache#readme",
   "dependencies": {
-    "@blackbaud/skyux-logger": "^1.0.2",
+    "@blackbaud/skyux-logger": "1.0.2",
     "cheerio": "0.22.0",
-    "cross-spawn": "^6.0.5",
+    "cross-spawn": "6.0.5",
     "fs-extra": "3.0.1",
     "glob": "7.1.2",
-    "latest-version": "^4.0.0",
+    "latest-version": "4.0.0",
     "marked": "0.3.9",
     "reserved-words": "0.1.1"
   },
